Crud customer added

+++ src/main/java/com/store/api/persistence/repository/ICustomerRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ICustomerRepository extends JpaRepository<Customer, Integer> {
+public interface ICustomerRepository extends JpaRepository<Customer, Long> {
 
     @Query("SELECT c FROM Customer c WHERE c.dni = :dni")
     Optional<Customer> findByDni(@Param("dni")String dni);
